requirements.txt
fast_bss_eval==0.1.4
fire==0.4.0
joblib==1.1.0
kaldiio==2.18.0
librosa==0.10.1
lmdb==1.3.0
matplotlib==3.5.1
mir_eval==0.7
numpy==1.21.5
pesq==0.0.4
pystoi==0.3.3
PyYAML==6.0
Requests==2.31.0
scipy==1.7.3
soundfile==0.12.1
tableprint==0.9.1
thop==0.1.1.post2209072238
torchnet==0.0.4
tqdm==4.64.0
flake8==3.8.2
flake8-bugbear
flake8-comprehensions
flake8-executable
flake8-pyi==20.5.0
auraloss
torchmetrics==1.2.0
h5py
